fix: increase grace period
The grace period was selected arbitrarily. It is the time the idle state monitor waits for new events to come in before reporting and idle state.

The value worked well initially, but lately we had an increasing number of flaky tests. This may be related to the synchronized code blocks that were introduced to improve multi-threading behavior

This hopefully fixes #119