ci: build and deploy engine as docker container
Added a job to the deployment pipeline which will build a container from the engine, and push it to dockerhub.

The container will default to the "latest" tag. If a new release is created in GitHub the name of the tag attached to this release will be used as the image tag, unless this release is a SNAPSHOT release.